pip_requirements.txt
aniso8601 @ file:///tmp/build/80754af9/aniso8601_1614708839015/work
blinker @ file:///C:/b/abs_b1i87khtob/croot/blinker_1737448732095/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
certifi @ file:///C:/b/abs_3beajm7umk/croot/certifi_1745939228545/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///C:/b/abs_42rbjngfys/croot/click_1744271606983/work
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1729036581634/work
Flask @ file:///C:/b/abs_a6m37uz578/croot/flask_1737454335662/work
Flask-RESTful @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/flask-restful_1729090733940/work
Flask-SQLAlchemy @ file:///C:/Users/dev-admin/buildout/croot/flask-sqlalchemy_1735870223488/work
greenlet @ file:///C:/b/abs_bbl_18a2tm/croot/greenlet_1733860081977/work
idna @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/idna_1729038997316/work
itsdangerous @ file:///C:/Users/dev-admin/buildout/croot/itsdangerous_1732929022589/work
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
mysql-connector-python @ file:///C:/Users/dev-admin/buildout/croot/mysql-connector-python_1735874113617/work/mysql-connector-python
protobuf @ file:///C:/b/abs_b5v84rwh56/croot/protobuf_1742419721750/work/bazel-bin/python/dist/protobuf-5.29.3-cp313-abi3-win_amd64.whl#sha256=0ef6aadfdcd9f988cbfc2d48eb8f7ed3d64acc158314c82a3a9c456f3fcc53a4
PyMySQL @ file:///C:/b/abs_3cmx0oymx2/croot/pymysql_1746768519259/work
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1729039320841/work
python-dateutil @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/python-dateutil_1729038406576/work
pytz @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pytz_1729039375330/work
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
setuptools==78.1.1
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
SQLAlchemy @ file:///C:/b/abs_275ybuj8n_/croot/sqlalchemy_1745854539419/work
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
Werkzeug @ file:///C:/b/abs_c7bupijx2_/croot/werkzeug_1737448709012/work
wheel==0.45.1
win_inet_pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1729038644686/work
